refactor: update theme application
