+++ .editorconfig
@@ -1,9 +1,9 @@
 root = true
 
 [*]
-end_of_line = lf
 indent_style = space
 indent_size = 2
+end_of_line = lf
 charset = utf-8
 trim_trailing_whitespace = true
 insert_final_newline = true
