+++ src/components/SocialShareButtons/SocialShareButtonsStyled.tsx
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import media from "styled-media-query";
 import Image from 'next/image'
 
 export const SocialShareContainer = styled.section`
@@ -7,9 +6,6 @@ export const SocialShareContainer = styled.section`
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  ${media.greaterThan("medium")`
-    align-items: flex-end;
-  `}
 `;
 
 export const SocialShareContent = styled.section`
@@ -20,7 +16,6 @@ export const SocialShareContent = styled.section`
 `;
 
 export const Title = styled.p`
-  color: ${(props) => props.theme.palette.text.secondary};
   line-height: 19px;
   text-align: center;
   max-width: 280px;
