--- src/styles/interfaces/Action.ts
@@ -1,13 +1,0 @@
-export interface IAction {
-  active?: string,
-  hover?: string,
-  hoverOpacity?: number | string,
-  selected?: string,
-  selectedOpacity?: number | string,
-  disabled?: string,
-  disabledBackground?: string,
-  disabledOpacity?: number | string,
-  focus?: string,
-  focusOpacity?: number | string,
-  activatedOpacity?: number | string,
-}
