--- src/styles/interfaces/Color.ts
@@ -1,6 +1,0 @@
-export interface IColor {
-  light?: string,
-  main: string,
-  dark?: string,
-  contrastText?: string,
-}
