--- src/styles/interfaces/Grey.ts
@@ -1,16 +1,0 @@
-export interface IGrey {
-  50: string,
-  100: string,
-  200: string,
-  300: string,
-  400: string,
-  500: string,
-  600: string,
-  700: string,
-  800: string,
-  900: string,
-  A100: string,
-  A200: string,
-  A400: string,
-  A700: string,
-}
