--- src/styles/interfaces/Margin.ts
@@ -1,5 +1,0 @@
-export interface IMargin {
-  small: string,
-  regular: string,
-  large: string,
-}
