--- src/styles/interfaces/Shape.ts
@@ -1,6 +1,0 @@
-export interface IShape {
-  borderRadius: number | string,
-  padding: {
-    small: string,
-  }
-}
