--- src/styles/interfaces/Text.ts
@@ -1,5 +1,0 @@
-export interface IText {
-  primary?: string,
-  secondary?: string,
-  disabled?: string,
-}
