--- src/styles/interfaces/Zindex.ts
@@ -1,9 +1,0 @@
-export interface IZindex {
-  mobileStepper?: number,
-  speedDial?: number,
-  appBar?: number,
-  drawer?: number,
-  modal?: number,
-  snackbar?: number,
-  tooltip?: number,
-}
