+++ src/styles/interfaces/index.ts
@@ -1,8 +1,1 @@
-export type { IColor } from "./Color";
-export type { IText } from "./Text";
-export type { IMargin } from "./Margin";
 export type { ITypography } from "./Typography";
-export type { IZindex } from "./Zindex";
-export type { IAction } from "./Action";
-export type { IShape } from "./Shape";
-export type { IGrey } from "./Grey";
